requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
future==1.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
jiter==0.9.0
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
openai==1.74.0
packaging==24.2
protobuf==6.30.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
websockets==15.0.1
